requirements.txt
virtualenv==20.0.31
python-dotenv==0.14.0
pytest==6.0.1
requests==2.24.0
discord.py==1.4.1